Separate all the component files, clean the project hierachy. Complete the project.
