--- 05-travel-list/src/App.js
@@ -1,176 +1,0 @@
-import { useState } from "react";
-
-// const initialItems = [
-//   { id: 1, description: "Passports", quantity: 2, packed: false },
-//   { id: 2, description: "Socks", quantity: 12, packed: false },
-//   { id: 3, description: "iPad", quantity: 1, packed: true },
-// ];
-
-export default function App() {
-  const [items, setItems] = useState([]);
-
-  function handleAddItems(item) {
-    setItems((items) => [...items, item]);
-  }
-
-  function handleDeleteItem(id) {
-    // filter out the item with the given id
-    setItems((items) => items.filter((item) => item.id !== id));
-  }
-
-  function handleToggleItem(id) {
-    setItems((items) =>
-      items.map((item) =>
-        item.id === id ? { ...item, packed: !item.packed } : item
-      )
-    );
-  }
-
-  function handleClearList() {
-    const confirmed = window.confirm(
-      "Are you sure you want to clear the list?"
-    );
-    if (!confirmed) return;
-    setItems([]);
-  }
-
-  return (
-    <div className="app">
-      <Logo />
-      <Form onAddItems={handleAddItems} />
-      <PackingList
-        items={items}
-        onDeleteItem={handleDeleteItem}
-        onToggleItem={handleToggleItem}
-        onClearList={handleClearList}
-      />
-      <Stats items={items} />
-    </div>
-  );
-}
-
-function Logo() {
-  return <h1 className="">🏝️FAR AWAY🥂</h1>;
-}
-
-function Form({ onAddItems }) {
-  const [description, setDescription] = useState("");
-  const [quantity, setQuantity] = useState(1);
-
-  function handleSubmit(e) {
-    e.preventDefault();
-
-    if (!description) return;
-
-    const newItem = { description, quantity, packed: true, id: Date.now() };
-
-    onAddItems(newItem);
-    // set description and quantity to the initial state
-    setDescription("");
-    setQuantity(1);
-  }
-
-  return (
-    <form className="add-form" onSubmit={handleSubmit}>
-      <h3>What do you need for your 😍 trip?</h3>
-      <select
-        value={quantity}
-        onChange={(e) => setQuantity(Number(e.target.value))}
-      >
-        {Array.from({ length: 20 }, (_, i) => i + 1).map((num) => (
-          <option key={num} value={num}>
-            {num}
-          </option>
-        ))}
-      </select>
-      <input
-        type="text"
-        placeholder="Item..."
-        value={description}
-        onChange={(e) => setDescription(e.target.value)}
-      />
-      <button>ADD</button>
-    </form>
-  );
-}
-
-function PackingList({ items, onDeleteItem, onToggleItem, onClearList }) {
-  const [sortBy, setSortBy] = useState("input");
-
-  let sortedItems;
-  if (sortBy === "input") sortedItems = items;
-
-  if (sortBy === "description") {
-    sortedItems = items.slice().sort((a, b) => {
-      return a.description.localeCompare(b.description);
-    });
-  }
-
-  if (sortBy === "status") {
-    sortedItems = items.slice().sort((a, b) => {
-      return Number(a.packed) - Number(b.packed);
-    });
-  }
-
-  return (
-    <div className="list">
-      <ul>
-        {sortedItems.map((item) => (
-          <Item
-            item={item}
-            onDeleteItem={onDeleteItem}
-            onToggleItem={onToggleItem}
-            key={item.id}
-          ></Item>
-        ))}
-      </ul>
-      <div className="actions">
-        <select value={sortBy} onChange={(e) => setSortBy(e.target.value)}>
-          <option value="input">Sort by input order</option>
-          <option value="status">Sort by status</option>
-          <option value="description">Sort by description</option>
-        </select>
-        <button onClick={onClearList}>Clear list</button>
-      </div>
-    </div>
-  );
-}
-
-function Item({ item, onDeleteItem, onToggleItem }) {
-  return (
-    <li>
-      <input
-        type="checkbox"
-        value={item.packed}
-        onChange={() => onToggleItem(item.id)}
-      />
-      <span style={item.packed ? {} : { textDecoration: "Line-through" }}>
-        {item.quantity} {item.description}
-      </span>
-      <button onClick={() => onDeleteItem(item.id)}>❌</button>
-    </li>
-  );
-}
-
-function Stats({ items }) {
-  if (!items.length)
-    return (
-      <p className="stats">
-        <em>Start adding some items to your list! 🚀</em>
-      </p>
-    );
-  const numItems = items.length;
-  const numPacked = items.filter((item) => !item.packed).length;
-  const percentage = Math.round((numPacked / numItems) * 100);
-  return (
-    <footer className="stats">
-      <em>
-        {percentage === 100
-          ? "You got everything ready! 🎉"
-          : `
-        You have ${numItems} items on your list, and you already packed
-        ${numPacked} items. That's ${percentage}%!`}
-      </em>
-    </footer>
-  );
-}
