+++ 05-travel-list/src/components/Stats.js
@@ -1,0 +1,22 @@
+export default function Stats({ items }) {
+  if (!items.length)
+    return (
+      <p className="stats">
+        <em>Start adding some items to your list! 🚀</em>
+      </p>
+    );
+  const numItems = items.length;
+  const numPacked = items.filter((item) => !item.packed).length;
+  const percentage = Math.round((numPacked / numItems) * 100);
+  return (
+    <footer className="stats">
+      <em>
+        {percentage === 100
+          ? "You got everything ready! 🎉"
+          : `
+          You have ${numItems} items on your list, and you already packed
+          ${numPacked} items. That's ${percentage}%!`}
+      </em>
+    </footer>
+  );
+}
